image-py/pyproject.toml
[tool.poetry]
name = "image-py"
version = "0.1.0"
description = ""
authors = ["moisutsu <moisutsu@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.21.1"
Pillow = "^8.3.1"
classopt = "^0.1.4"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^21.7b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
